tools/release/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
docker==7.0.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
packaging==23.2
requests==2.31.0
smmap==5.0.1
termcolor==2.4.0
urllib3==2.2.0
